site/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
astroid==1.4.9
awesome-slugify==1.6.5
billiard==3.3.0.23
celery==3.1.25
crc16==0.1.1
Django==1.11
django-angular==0.8.4
django-appconf==1.0.2
django-celery-email==1.1.5
django-ckeditor==5.2.1
django-configurations==2.0
django-debug-toolbar==1.7
djangorestframework==3.6.2
easy-thumbnails==2.3
gunicorn==19.7.1
isort==4.2.5
kombu==3.0.37
lazy-object-proxy==1.2.2
mccabe==0.6.1
mysqlclient==1.3.10
olefile==0.44
packaging==16.8
Pillow==4.1.0
ptvsd==3.0.0
pylint==1.6.5
pyparsing==2.2.0
python-redis==0.1.4
pytz==2017.2
regex==2017.2.8
six==1.10.0
sqlparse==0.2.3
Unidecode==0.4.20
wrapt==1.10.10
django-celery-results
django-bootstrap-pagination
django-breadcrumbs

